requirements.txt
# LLMBroker 依赖包

# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# HTTP客户端
httpx==0.25.2

# 配置文件解析
PyYAML==6.0.1

# 类型注解支持
typing-extensions==4.8.0

# 异步支持
anyio==4.1.0

# 数据验证
pydantic==2.5.0

# 日志处理
loguru==0.7.2

# 环境变量管理
python-dotenv==1.0.0

# 性能监控（可选）
prometheus-client==0.19.0

# 开发依赖
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# 代码格式化
black==23.11.0
isort==5.12.0

# 代码质量检查
flake8==6.1.0
mypy==1.7.1
